requirements.txt
altgraph==0.17.3
click==8.1.3
colorama==0.4.6
Flask==2.2.2
future==0.18.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.21.6
opencv-python-headless==4.7.0.68
pefile==2022.5.30
Pillow==9.4.0
psycopg2==2.9.4
pyinstaller==5.7.0
pyinstaller-hooks-contrib==2022.14
pywin32-ctypes==0.2.0
Werkzeug==2.2.2
xlrd==1.2.0
